Remove SVM's dependency on system program
