+++ programs/system/Cargo.toml
@@ -35,6 +35,7 @@ solana-compute-budget = { workspace = true }
 solana-feature-set = { workspace = true }
 solana-hash = { workspace = true }
 solana-logger = { workspace = true }
+solana-nonce-account = { workspace = true }
 solana-rent = { workspace = true }
 solana-sdk = { workspace = true }
 solana-sha256-hasher = { workspace = true }
