+++ runtime/Cargo.toml
@@ -67,6 +67,7 @@ solana-inline-spl = { workspace = true }
 solana-lattice-hash = { workspace = true }
 solana-measure = { workspace = true }
 solana-metrics = { workspace = true }
+solana-nonce-account = { workspace = true }
 solana-perf = { workspace = true }
 solana-program = { workspace = true }
 solana-program-runtime = { workspace = true, features = ["metrics"] }
